Update
Added an example folder to show off the cors error when sending a request from the frontend. To use it, you'd spin up live server on the contained index.html file, or open the file path in your browser. I also changed some of the variable naming and target ID in public/index.html, just for clarity.